+++ src/tuntap/tun_darwin.go
@@ -16,7 +16,7 @@ import (
 )
 
 // Configures the "utun" adapter with the correct IPv6 address and MTU.
-func (tun *TunAdapter) setup(ifname string, addr string, mtu uint16) error {
+func (tun *TunAdapter) setup(ifname string, addr string, mtu uint64) error {
 	if ifname == "auto" {
 		ifname = "utun"
 	}
@@ -26,7 +26,7 @@ func (tun *TunAdapter) setup(ifname string, addr string, mtu uint16) error {
 	}
 	tun.iface = iface
 	if m, err := iface.MTU(); err == nil {
-		tun.mtu = getSupportedMTU(uint16(m))
+		tun.mtu = getSupportedMTU(uint64(m))
 	} else {
 		tun.mtu = 0
 	}
